+++ src/components/playpadMode/PlayPad.js
@@ -1,8 +1,8 @@
 import React, { useState, useContext, useEffect } from "react";
 import * as Tone from "tone";
 import "./PlayPad.css";
-import { presetBankData } from "./PresetDesign";
-import { playTargetPlayer, quantizeTransportPosition } from "./helperFunctions";
+import { presetBankData } from "../PresetDesign";
+import { playTargetPlayer, quantizeTransportPosition } from "../../helperFunctions";
 
 function PlayPad({ padIndex, padColor, localRecs, setLocalRecs }) {
   //unpack data layer values
