+++ src/components/sequenceMode/Sequencer.css
