+++ src/components/sequenceMode/SequencerItem.css
