+++ src/components/sequenceMode/SequencerRow.css
