+++ src/components/sequenceMode/SequencerRow.js
