+++ src/reducers/isActiveReducer.js
